requirements.txt
pyserial

PySimpleGUI


pycodestyle
pydocstyle
pylint

# ipykernel
